# docs/_configuration/openapi.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "/",
      "description": "Default Server URL"
    }
  ],
  "paths": {
...
    "/admin/api/v1/eventstreams/{collectionName}/views": {
      "get": {
        "tags": [
          "Views"
        ],
        "summary": "Retrieve a list of configured views of a collection",
        "operationId": "getViews",
        "parameters": [
          {
            "name": "collectionName",
            "in": "path",
            "description": "The name of the collection",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "event-stream"
          }
        ],
        "responses": {
          "404": {
            "description": "Missing EventStream"
          },
          "200": {
            "description": "OK",
            "content": {
              "text/turtle": {},
              "application/ld+json": {},
              "application/n-quads": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "Views"
        ],
        "summary": "Add view to a collection",
        "operationId": "createView",
        "parameters": [
          {
            "name": "collectionName",
            "in": "path",
            "description": "The name of the collection",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "event-stream"
          }
        ],
        "requestBody": {
          "description": "A valid RDF model defining a view for a collection",
          "content": {
            "text/turtle": {
              "schema": {
                "type": "string"
              },
              "examples": {
                "Time-Based Retention": {
                  "description": "A time-based retention policy which is configured to only keep members whose ldes:timestamppath is less than 2 minutes ago.",
                  "value": "@prefix ldes: <https://w3id.org/ldes#> .\n@prefix tree: <https://w3id.org/tree#> .\n@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .\n@prefix server: <http://localhost:8080/event-stream/> .\n\nserver:time-based-retention tree:viewDescription [\n\ta tree:fragmentationStrategy;\n\ttree:fragmentationStrategy  () ;\n    tree:pageSize \"100\"^^<http://www.w3.org/2001/XMLSchema#int> ;\n    ldes:retentionPolicy [\n        a ldes:DurationAgoPolicy  ;\n        tree:value \"PT2M\"^^xsd:duration ;\n    ] ;\n] .\n"
                },
                "Version-Based Retention": {
                  "description": "A version-based retention policy which is configured to only keep members the two most recent versions of a resource.",
                  "value": "@prefix ldes: <https://w3id.org/ldes#> .\n@prefix tree: <https://w3id.org/tree#>.\n@prefix xsd: <http://www.w3.org/2001/XMLSchema#> .\n@prefix server: <http://localhost:8080/event-stream/> .\n\nserver:version-based-retention tree:viewDescription [\n\ttree:fragmentationStrategy  () ;\n    tree:pageSize \"100\"^^<http://www.w3.org/2001/XMLSchema#int> ;\n    ldes:retentionPolicy [\n        a ldes:LatestVersionSubset;\n        ldes:amount 2 ;\n    ] ;\n] .\n"
                },
                "Pagination View": {
                  "description": "A pagination fragmentation strategy which is configured to create new pages when a member limit of 100 members is reached.",
                  "value": "@prefix server: <http://localhost:8080/event-stream/> .\n@prefix tree: <https://w3id.org/tree#> .\n\nserver:pagination tree:viewDescription [\n\ttree:fragmentationStrategy  () ;\n    tree:pageSize \"100\"^^<http://www.w3.org/2001/XMLSchema#int> ;\n ] .\n"
                },
                "Time-Based Fragmentation Strategy": {
                  "description": "A time-based fragmentation strategy which is configured to create new pages when a member limit of 100 members is reached.",
                  "value": "@prefix server: <http://localhost:8080/event-stream/> .\n@prefix tree: <https://w3id.org/tree#> .\n@prefix ldes: <https://w3id.org/ldes#> .\n\nserver:timebased tree:viewDescription [\n\ttree:fragmentationStrategy ([\n\t\ta tree:HierarchicalTimeBasedFragmentation ;\n\t\ttree:maxGranularity \"day\" ;\n\t\ttree:linearTimeCachingEnabled \"true\" ;\n        tree:fragmentationPath ldes:timestampPath ;\n    ]) ;\n    tree:pageSize \"100\"^^<http://www.w3.org/2001/XMLSchema#int> ;\n ] .\n"
                },
                "Geospatial Fragmentation Strategy": {
                  "description": "A geospatial fragmentation strategy which is configured to partition the data on zoom level 15.",
                  "value": "@prefix server: <http://localhost:8080/event-stream/> .\n@prefix tree: <https://w3id.org/tree#> .\n\nserver:geospatial tree:viewDescription [\n\ttree:fragmentationStrategy ([\n\t\ta tree:GeospatialFragmentation ;\n\t\ttree:maxZoom 15 ;\n\t\ttree:fragmentationPath <http://www.opengis.net/ont/geosparql#asWKT>\n\t]) ;\n\ttree:pageSize \"100\"^^<http://www.w3.org/2001/XMLSchema#int> ;\n] .\n"
                },
                "Reference Fragmentation Strategy": {
                  "description": "A reference fragmentation strategy which is configured to partition the data on the provided property path (fragmentationPath).",
                  "value": "@prefix server: <http://localhost:8080/event-stream/> .\n@prefix tree: <https://w3id.org/tree#> .\n\nserver:reference tree:viewDescription [\n\ttree:fragmentationStrategy ([\n\t\ta tree:ReferenceFragmentation ;\n\t\ttree:fragmentationPath <http://purl.org/dc/terms/isVersionOf> ;\n\t\ttree:fragmentationKey \"version\" ;\n\t]) ;\n\ttree:pageSize \"100\"^^<http://www.w3.org/2001/XMLSchema#int> ;\n] .\n"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Duplicate View or Wrongly Configured View"
          },
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Missing EventStream"
          }
        }
      }
    },
...
    "/admin/api/v1/eventstreams/{collectionName}/views/{viewName}": {
      "get": {
        "tags": [
          "Views"
        ],
        "summary": "Retrieve a specific view of a collection",
        "operationId": "getViewOfCollection",
        "parameters": [
          {
            "name": "collectionName",
            "in": "path",
            "description": "The name of the collection",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "event-stream"
          },
          {
            "name": "viewName",
            "in": "path",
            "description": "The name of requested view",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "time-based-retention"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/turtle": {},
              "application/ld+json": {},
              "application/n-quads": {}
            }
          },
          "404": {
            "description": "Missing EventStream or Missing View"
          }
        }
      }
    },
    "/admin/api/v1/eventstreams/{collectionName}/views/{view}": {
      "delete": {
        "tags": [
          "Views"
        ],
        "summary": "Delete a specific view for a collection",
        "operationId": "deleteView",
        "parameters": [
          {
            "name": "collectionName",
            "in": "path",
            "description": "The name of the collection",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "event-stream"
          },
          {
            "name": "view",
            "in": "path",
            "description": "The name of deleted view",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "time-based-retention"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/turtle": {},
              "application/ld+json": {},
              "application/n-quads": {}
            }
          },
          "404": {
            "description": "Missing EventStream or Missing View"
          }
        }
      }
    }
  },
...
}